+++ item-system/demonlord.js
@@ -1,8 +1,5 @@
+// Support for the Shadow of the Demon Lord system was initially added by Xacus
 mergeObject(CONFIG, {
-  /*
-  Support for the Shadow of the Demon Lord system was initially added by Xacus
-  */
-
   illandril: {
     hotbarUses: {
       macros: {
@@ -116,7 +113,7 @@ function calculateUsesForItem(item) {
 function calculateTalentUses(item) {
   const uses = item.uses;
   if (uses && (uses.max > 0 || uses.value > 0)) {
-    return uses.value + "/" + uses.max;
+    return uses.value + '/' + uses.max;
   }
 
   return null;
@@ -125,7 +122,7 @@ function calculateTalentUses(item) {
 function calculateSpellUses(item) {
   const castings = item.castings;
   if (castings && (castings.max > 0 || castings.value > 0)) {
-    return castings.value + "/" + castings.max;
+    return castings.value + '/' + castings.max;
   }
 
   return null;